moved the command out of the rockpaperscissors package and deleted that package
