+++ src/main/java/com/technovision/technobot/commands/CommandRegistry.java
@@ -4,7 +4,7 @@ import com.technovision.technobot.TechnoBot;
 import com.technovision.technobot.commands.economy.*;
 import com.technovision.technobot.commands.fun.CommandRoll;
 import com.technovision.technobot.commands.fun.CommandSay;
-import com.technovision.technobot.commands.fun.rockpaperscissors.CommandRockPaperScissors;
+import com.technovision.technobot.commands.fun.CommandRockPaperScissors;
 import com.technovision.technobot.commands.levels.CommandLeaderboard;
 import com.technovision.technobot.commands.levels.CommandRank;
 import com.technovision.technobot.commands.levels.CommandRankcard;
